requirements.txt
psycopg2-binary>=2.9.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
voyageai>=0.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
